requirements.txt
# line bot api ------------------------------------
line-bot-sdk<3.00
flask
werkzeug>=0.15.4
python-dotenv
gunicorn

# yolo ------------------------------------
Cython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.1
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0
thop

# plotting ------------------------------------
pandas>=1.1.4
seaborn>=0.11.0